making unknownKeywords.json schemas load properly
Now during initial $anchor lookup we do not traverse unhandled keywords.

It works as follows: SchemaLoader#lookupAnchors() examines the json pointer locating the current JsonValue, and continues the
recursion only if
 * the last segment of the pointer is a known json schema keyword
 * or the last-before segment is a known keyword, and it is known to have map-like semantics, like "properties" or "$defs"

To make this reasonably testable, the schema jsons being read from the json test files need adjustment as well: we have to trim
the leading <arrayindex>/"schema" (2 segments) of the schema json, so that the root schema object truly has an empty json pointer
instead of a 2-element pointer (which would not pass the above two criteria, therefore none of the $anchors would be recognized
during testing).