Refactor: per-keyword source files (#14)
Reorganizes the source code so that for each schema keyword the related `Schema` and `ValidationFailure` subclasses, and a small function used to load a schema json to the specific `Schema`. These small loaders are used by the `SchemaLoader`.

This refactor makes the source organization better because the previous `ValidationFailure.kt` and `Schema.kt` just grew too big.